+++ Monolith/app/controller/common.go
@@ -1,0 +1,6 @@
+package controller
+
+type Response struct {
+	StatusCode int32  `json:"status_code"`
+	StatusMsg  string `json:"status_msg"`
+}
