+++ Monolith/main.go
@@ -1,13 +1,15 @@
 package main
 
 import (
+	"mini-tiktok/router"
+
 	"github.com/gin-gonic/gin"
 )
 
 func main() {
 	r := gin.Default()
 
-	initRouter(r)
+	router.InitRouter(r)
 
 	r.Run() // listen and serve on 0.0.0.0:8080 (for windows "localhost:8080")
 }
