--- Monolith/router.go
@@ -1,34 +1,0 @@
-package main
-
-import (
-	"mini-tiktok/controller"
-	"mini-tiktok/middleware"
-
-	"github.com/gin-gonic/gin"
-)
-
-func initRouter(r *gin.Engine) {
-	// public directory is used to serve static resources
-	r.Static("/static", "./public")
-
-	apiRouter := r.Group("/douyin")
-
-	// basic apis
-	// apiRouter.GET("/feed/", controller.Feed)
-	apiRouter.GET("/user/", middleware.UserAuthMiddleware("Query"), controller.UserInfo)
-	apiRouter.POST("/user/register/", controller.Register)
-	apiRouter.POST("/user/login/", controller.Login)
-	// apiRouter.POST("/publish/action/", controller.Publish)
-	// apiRouter.GET("/publish/list/", controller.PublishList)
-
-	// extra apis - I
-	// apiRouter.POST("/favorite/action/", controller.FavoriteAction)
-	// apiRouter.GET("/favorite/list/", controller.FavoriteList)
-	// apiRouter.POST("/comment/action/", controller.CommentAction)
-	// apiRouter.GET("/comment/list/", controller.CommentList)
-
-	// extra apis - II
-	apiRouter.POST("/relation/action/", middleware.UserAuthMiddleware("Query"), controller.RelationAction)
-	// apiRouter.GET("/relation/follow/list/", controller.FollowList)
-	// apiRouter.GET("/relation/follower/list/", controller.FollowerList)
-}
